add qmls in resources and update main.py startup.py new_project.py
