+++ src/main/python/main.py
@@ -1,6 +1,6 @@
 from PyQt5 import Qt, QtCore
 from PyQt5.QtWidgets import QStyleFactory
-from view.startup import StartupWidget
+from view.startup import StartupEngine
 from model.fbs import AppInfo
 from fbs_runtime.application_context import cached_property, is_frozen
 from fbs_runtime.application_context.PyQt5 import ApplicationContext
@@ -36,23 +36,22 @@ class AppContext(ApplicationContext):           # 1. Subclass ApplicationContext
         scope.set_extra('build', AppInfo().version())
 
     def run(self):                             # 2. Implement run()
-        """ start QtApplication """
-        startup_window = StartupWidget()
-        startup_window.setWindowTitle(AppInfo().app_name() + ' Version ' + AppInfo().version())
+        """ start QQmlApplicationEngine """
         if len(sys.argv) > 1 and type(sys.argv[1]) is str:
             project_file_path = sys.argv[1]
             _, ext = os.path.splitext(project_file_path)
             if ext == '.sdt':
-                startup_window.move_to_main_window(project_file_path)
+                # startup_window.move_to_main_window(project_file_path)
+                engine = StartupEngine()
         else:
-            startup_window.show()
+            engine = StartupEngine()
 
         # スタイルをwindows共用に(for develop)
         # self.app.setStyle(QStyleFactory.create('Fusion'))
 
-        # Enable High DPI display with PyQt5
-        os.environ["QT_AUTO_SCREEN_SCALE_FACTOR"] = "1"
-        self.app.setAttribute(QtCore.Qt.AA_EnableHighDpiScaling, True)
+        # TODO: Enable High DPI display with PyQt5 for QML
+        # os.environ["QT_AUTO_SCREEN_SCALE_FACTOR"] = "1"
+        # self.app.setAttribute(QtCore.Qt.AA_EnableHighDpiScaling, True)
 
         return self.app.exec_()
